removed svg minifier and moved its file because of wasm compiler issues
